refactor: move constructor to mixins
