+++ src/components/core/common-behaviors/decorators/host-attributes.ts
@@ -1,6 +1,6 @@
 import { isServer, type ReactiveElement } from 'lit';
 
-import type { AbstractConstructor } from '../constructor';
+import type { AbstractConstructor } from '../mixins/constructor';
 
 function applyAttributes(instance: ReactiveElement, attributes: Record<string, string>): void {
   for (const [name, value] of Object.entries(attributes)) {
