+++ src/components/core/common-behaviors/index.ts
@@ -1,5 +1,5 @@
 export * from './base-classes';
-export * from './constructor';
+export * from './mixins/constructor';
 export * from './controllers';
 export * from './decorators';
 export * from './mixins';
