+++ src/components/core/common-behaviors/mixins/negative-mixin.ts
@@ -1,7 +1,7 @@
 import type { LitElement } from 'lit';
 import { property } from 'lit/decorators.js';
 
-import type { AbstractConstructor } from '../constructor';
+import type { AbstractConstructor } from './constructor';
 
 export declare class SbbNegativeMixinType {
   public negative: boolean;
