+++ src/components/core/common-behaviors/mixins/update-scheduler-mixin.ts
@@ -1,6 +1,6 @@
 import type { LitElement } from 'lit';
 
-import type { Constructor } from '../constructor';
+import type { Constructor } from './constructor';
 
 // Define the interface for the mixin
 export declare class UpdateSchedulerMixinType {
